requirements.txt
streamlit>=1.37.0
flask>=3.0.0
opencv-python>=4.10.0.84
pdf2image>=1.17.0
Pillow>=10.4.0
numpy>=1.26.0
pandas>=2.0.0
openpyxl>=3.1.0

